Update and rename configManager.cs to ConfigManager.cs
